Revert "На нахуй сука черти ебанные я пошел хавать блядь охуенный тгстейшн"
This reverts commit 1eec49beea6113a3791a67a22464ccbab1b6979b.

+++ code/modules/spells/spell_types/touch/smite.dm
@@ -1,7 +1,7 @@
 /datum/action/cooldown/spell/touch/smite
-	name = "Smite"
-	desc = "This spell charges your hand with an unholy energy \
-		that can be used to cause a touched victim to violently explode."
+	name = "Кара"
+	desc = "Заклинание заряжает вашу руку нечестивой энергией. \
+		Если вы дотронетесь ею до жертвы, то её тело взорвется в кровавом месиве.."
 	button_icon_state = "gib"
 	sound = 'sound/magic/disintegrate.ogg'
 
@@ -14,46 +14,42 @@
 
 	hand_path = /obj/item/melee/touch_attack/smite
 
-/// Smite is pretty extravagant, so whenever we get casted, we blind everyone nearby.
-/datum/action/cooldown/spell/touch/smite/proc/blind_everyone_nearby(mob/living/victim, atom/center)
+/datum/action/cooldown/spell/touch/smite/cast_on_hand_hit(obj/item/melee/touch_attack/hand, atom/victim, mob/living/carbon/caster)
+	if(!isliving(victim))
+		return FALSE
+
 	do_sparks(sparks_amt, FALSE, get_turf(victim))
-	for(var/mob/living/nearby_spectator in view(center, 7))
-		if(nearby_spectator == center)
+	for(var/mob/living/nearby_spectator in view(caster, 7))
+		if(nearby_spectator == caster)
 			continue
 		nearby_spectator.flash_act(affect_silicon = FALSE)
 
-/datum/action/cooldown/spell/touch/smite/on_antimagic_triggered(obj/item/melee/touch_attack/hand, mob/living/victim, mob/living/carbon/caster)
-	caster.visible_message(
-		span_warning("The feedback blows [caster]'s arm off!"),
-		span_userdanger("The spell bounces from [victim]'s skin back into your arm!"),
-	)
-	// Off goes the arm we were casting with!
-	var/obj/item/bodypart/to_dismember = caster.get_holding_bodypart_of_item(hand)
-	to_dismember?.dismember()
-	// And do the blind (us included)
-	caster.flash_act()
-	blind_everyone_nearby(caster, caster)
-
-/datum/action/cooldown/spell/touch/smite/cast_on_hand_hit(obj/item/melee/touch_attack/hand, mob/living/victim, mob/living/carbon/caster)
-	blind_everyone_nearby(victim, caster)
+	var/mob/living/living_victim = victim
+	if(living_victim.can_block_magic(antimagic_flags))
+		caster.visible_message(
+			span_warning("Обратная связь отрывает руку [caster]!"),
+			span_userdanger("Заклинание отскакивает от кожи [living_victim] обратно в мою руку!"),
+		)
+		caster.flash_act()
+		var/obj/item/bodypart/to_dismember = caster.get_holding_bodypart_of_item(hand)
+		to_dismember?.dismember()
+		return TRUE
 
 	if(ishuman(victim))
 		var/mob/living/carbon/human/human_victim = victim
 		var/obj/item/clothing/suit/worn_suit = human_victim.wear_suit
 		if(istype(worn_suit, /obj/item/clothing/suit/hooded/bloated_human))
-			human_victim.visible_message(span_danger("[victim]'s [worn_suit] explodes off of them into a puddle of gore!"))
+			human_victim.visible_message(span_danger("[victim] [worn_suit] взрывается, превращаясь в лужу запекшейся крови!"))
 			human_victim.dropItemToGround(worn_suit)
 			qdel(worn_suit)
 			new /obj/effect/gibspawner(get_turf(victim))
 			return TRUE
 
-	victim.investigate_log("has been gibbed by the smite spell.", INVESTIGATE_DEATHS)
-	victim.gib(DROP_ALL_REMAINS)
+	living_victim.gib()
 	return TRUE
 
 /obj/item/melee/touch_attack/smite
-	name = "\improper smiting touch"
-	desc = "This hand of mine glows with an awesome power!"
-	icon = 'icons/obj/weapons/hand.dmi'
+	name = "карающее прикосновение"
+	desc = "Моя рука светится невероятной силой!"
 	icon_state = "disintegrate"
 	inhand_icon_state = "disintegrate"
